FFM-1825 Auto target registration
- Creates a client service that we can use to forward auth requests to
- Adds logic in the ProxyServices Authenticate method to add Targets to
  its cache and then forward the Auth request on to the Client Service
which will then add the Target to feature flags.

Testing
- Added two test cases for the Authentication endpoint so that when a
  target is included in the paylaod we can verify that it will be added
to the cache and client service (provided we've a working connection
with the client service). I had to tweak the setupServer function
slightly so I could pass in a TargetRepo and mock ClientService so I
could assert that the Targets were being stored in t