FFM-8999 Prevent potential panic in metrics client
**What**

- Adds a nil pointer check

**Why**

- We were seeing some panics in the logs in QA for the line that
  attempted to append TargetData for the incoming request to the
currentRequest.TargetData slice. I suspect this was happening because
we'd get a metrics request for an environmnet with no TargetData, then
get a second request for the same environment with TargetData. Because
there was no TargetData in the first request that meant
`currentMetrics.TargetData` was nil and we panicked.

**Testing**

I was able to recreate panics through the two unit tests I added so
added a nil check to the code.