FFM-9893 Refresh Token before fetching new env config
**What**

- Forces the Proxy to refresh the auth token it uses to fetch env config

**Why**

- When the Proxy starts up it authenticates and stores an auth token
  that it uses for future requests. The issue with this though is the
auth token is only authorised for use with environments that existed
when the Proxy started up. When we add a new environment and the Proxy
tries to fetch this config it gets an unauthorised error because the
toekn its using doesn't have access to the environmnet. To fix this we
refresh the auth token which means the new one will have access to the
newly created environment

**Testing**

- Tested manually locally