FFM-10697 Fix stuck in INITIALIZING stream STATE
**What**

- Stores the StreamStatus in memory as well as in the redis cache. The
  in memory status can't be impacted by network issues with redis so
should always be correct and up to date.
- Starts a goroutine that periodically checks the cached StreamStatus
  matches the in memory StreamStatus

**Why**

There was a bug where the Primary Proxy could start up, succesfully
stream with Saas and fail to update the StreamState in redis from
INTIALIZING to CONNECTED due to a network error. This would then lead to
the StreamState being stuck at INITIALIZING which meant SDK /stream
requests were rejected by the Replica Proxy's. The Proxy would then end
up stuck in this state until there was a genuine disconnect and
reconnect with the Saas stream which would then correct the StreamState
in redis.

**Testing**

- I manually commented out the code that sets the stream state from
  INITIALIZING to CONNECTED at startup and saw that when the ticker in
the Gorotuine expired the StreamStatus was corrected to match the
inMemoryState.