FFM-10697 Fix attribtue evaluation bug
**What**

- Updates the version of the GoSDK that the Proxy uses. This change was
  made to V1 but we missed porting it over during development of V2
- Fixes a bug with the Flag marshaling to the SDK type. We were setting
  the Distribution field during the conversion even when it was nil in
the original object. This then caused a bug in client side evaluations
when targeting attributes.

**Why**

- When I pointed my SDk at Saas I would get the correct evaluation based
  on target attribute. But when I pointed it at the Proxy it was always
falling back to the default evaluation and didn't seem to be taking in
to account the attribute rule

**Testing**

- Locally I commented out the old/new code to swap between them and with
  the new code the Proxy behaves the same as Saas.
- Wrote a unit test and ran it agains the old and new code. When I ran
  it against the old code it failed because we converted nil to a
default Distribution struct. When I ran it with the new code is passes
meaning we leave nil as nil and don't convert to a default struct