FFM-10731  Fix bug where bad stream state could be recored with redis connection issues
**What**

- We only update the in memory stream status if we don't error fetching
  the cached status from redis

**Why**

- At the start of the SetHealthy & SetUnhealthy functions we decalre an
  empty StreamStatus. We then have a defer that sets that variable as
the inMemoryStatus at the end of the function. However if we got an
InternalError getting the cached value from redis which happened if
there was a connection issue, then we'd end up
setting a default struct with empty fields as the in memory stream
state.

This resulted in a bug where the stream state could be this in the
health response
```
{"configStatus":{"state":"SYNCED","since":1708009070054},"streamStatus":{"state":"","since":0},"cacheStatus":"healthy"}
```

Instead of being `INITIALIZING`, `CONNECTED` or `DISCONNECTED`

**Testing**

- Added unit tests that mock out the redis connection error