FFM-11310 Harden StreamStatus (#298)
* FFM-11310 Harden StreamStatus

**What**

- Changes how the replica is notified of stream disconnects and keeps
 its in memory stream status in sync. With the old code the replica
polled redis to get the stream status which could result in a lag. Now
the replica hits redis at startup to get the stream status, but from
that point on relies on the disconnect and connect messages that the
Primary sends it.

- Creates seperate Health types for the Primary and Replica. Previously
  we'd the one type used for both and there was something dodgy going on
when it ran in a read replica proxy. The seperate types should make
debugging a lot easier.

**Why**

- We've seen some hard to reproduce buggy behaviour where the read
  replica status can get stuck on DISCONNECTED

**Testing**

- Added additional unit tests
- Running a Primary & Replica locally with a curl script checking their
  health endpoints and they've been staying in sync