FFM-12445 - Wait for pushpin to be available before starting up
**What**

- Makes the Proxy wait for a pushpin healthcheck to pass before starting
  up

**Why**

- We seem to occassionally get a bunch of error logs around startup
  where we're trying to publish events to pushpin but get a connection
refused error. Because this is happening at startup I suspect what might
be happening is we're trying to publish events to pushpin before it's
ready to receving them. This should make us wait until we're able to
publish events to pushpin before we start to try and sending real SSE events.

**Testing**

- Manually tested locally by taking down pushpin and I can replicate the
  connection refused error we see in our logs. Then checked that this
makes us wait properly until pushpin is back up.